demos: high_res: Cleanly exit for all exit modes
The `exit_cb()` function handles the clearing of the screen when the
demo is closed through the "log out" button on the UI. This function
frees up the memory used by the demo, clears the framebuffer etc.

Presently, this function does not handle signals such as SIGINT
(triggered by the user pressing Crtl-C) and SIGTERM (triggered by
systemd's "stop" command). As a result, the memory isn't freed and the
framebuffer still shows the last screen even after the demo has exited.

Make `exit_cb()` the handler for these signals, so that we can have a
clean exit from the demo in all situations.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>