demos: high_res: ADC: Exit the thread if ADC doesn't exist
Major features of the demo, including the ADC, are handled by separate
threads. In ADC's case, if the ADC does not exist on the SoC (or, if it
exists but has not been enabled), the ADC thread continues to run and
repeatedly attempts to open ADC's sysfs entry, thereby polluting the
console and wasting system resources.

To fix this, check at the beginning of the `adc_init()` function (called
by `pthread_create()`) if the ADC is available. If it isn't available,
exit the thread.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>