demos: high_res: Call `pclose()` after `popen()`
There are 2 related issues observed in TI LVGL Demo, and both of these
occur after running the demo for a while:
    1. The date stops showing correctly
    2. ADC value stops updating

This happens because we aren't calling `pclose()` inside the
ever-running loop in the ADC thread. This leads to too many files being
opened (as reported by perror logs). So the system stops entertaining
more `open file` requests, leading to operations depended on opening
files (such as ADC and date) to malfunction.

Fix this by calling `pclose()` in ADC thread, inside the loop. Therefore
now only one popen file descriptor is active at a given iteration.

Also call `pclose()` in the button thread, since it is missing there
too.

For testing, I ran the demo built with this patch for 18 hours on AM62L,
and both date and ADC value continued to show correctly.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>